demanding clients

+++ src/main/java/progetto/bigdata/sparkjobexecutor/query/DemandingClientsPerNationality.java
@@ -9,7 +9,7 @@ public enum DemandingClientsPerNationality implements Query<CoppiaNazionalitàCl
         public CoppiaNazionalitàClientePunteggioMedio build(String line) {
             String s = line.substring(1,line.length()-1);
             String[] tokens = s.split(",");
-            return new CoppiaNazionalitàClientePunteggioMedio(tokens[0],Float.valueOf(tokens[1]));
+            return new CoppiaNazionalitàClientePunteggioMedio(tokens[0],Double.valueOf(tokens[1]));
         }
     }
 }
